Fixed CS and renamed constants class
